adds md5sums for segments as default feature. fix set_file_length()

+++ parcel/utils.py
@@ -1,8 +1,12 @@
 from progressbar import ProgressBar, Percentage, Bar, ETA, FileTransferSpeed
 
+import os
 from const import GB
 from log import get_logger
 import requests
+import hashlib
+import mmap
+from contextlib import contextmanager
 
 # Logging
 log = get_logger('utils')
@@ -48,10 +52,21 @@ def write_offset(path, data, offset):
     f.close()
 
 
+def read_offset(path, offset, size):
+    f = open(path, 'r+b')
+    f.seek(offset)
+    data = f.read(size)
+    f.close()
+    return data
+
+
 def set_file_length(path, length):
+    if os.path.isfile(path) and os.path.getsize(path) == length:
+        return
     f = open(path, 'wb')
     f.seek(length-1)
     f.write('\0')
+    f.truncate()
     f.close()
 
 
@@ -61,3 +76,16 @@ def calculate_segments(start, stop, block):
 
     """
     return [(a, min(stop, a+block)-1) for a in range(start, stop, block)]
+
+
+def md5sum(block):
+    m = hashlib.md5()
+    m.update(block)
+    return m.hexdigest()
+
+
+@contextmanager
+def mmap_open(path):
+    with open(path, "r+b") as f:
+        mm = mmap.mmap(f.fileno(), 0)
+        yield mm
